Refactor main.rb file
